Adding basic dummy event
